DRY Sierra test code
