+++ sierra_adapter/sierra_item_merger/src/test/scala/uk/ac/wellcome/platform/sierra_item_merger/links/ItemUnlinkerTests.scala
@@ -3,8 +3,9 @@ package uk.ac.wellcome.platform.sierra_item_merger.links
 import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.models.transformable.SierraTransformable
 import uk.ac.wellcome.models.transformable.sierra.SierraItemRecord
+import uk.ac.wellcome.test.utils.SierraData
 
-class ItemUnlinkerTests extends FunSpec with Matchers {
+class ItemUnlinkerTests extends FunSpec with Matchers with SierraData {
 
   it("removes the item if it already exists") {
     val bibId = "222"
@@ -127,57 +128,4 @@ class ItemUnlinkerTests extends FunSpec with Matchers {
 
     caught.getMessage shouldEqual "Non-matching bib id 222 in item unlink bibs List(846)"
   }
-
-  def sierraItemRecord(
-    id: String = "i111",
-    title: String = "Ingenious imps invent invasive implements",
-    modifiedDate: String = "2001-01-01T01:01:01Z",
-    bibIds: List[String],
-    unlinkedBibIds: List[String] = List()
-  ) = SierraItemRecord(
-    id = id,
-    data = sierraRecordString(
-      id = id,
-      updatedDate = modifiedDate,
-      title = title
-    ),
-    modifiedDate = modifiedDate,
-    bibIds = bibIds,
-    unlinkedBibIds = unlinkedBibIds
-  )
-
-  private def sierraRecordString(
-    id: String,
-    updatedDate: String,
-    title: String
-  ) =
-    s"""
-       |{
-       |      "id": "$id",
-       |      "updatedDate": "$updatedDate",
-       |      "createdDate": "1999-11-01T16:36:51Z",
-       |      "deleted": false,
-       |      "suppressed": false,
-       |      "lang": {
-       |        "code": "ger",
-       |        "name": "German"
-       |      },
-       |      "title": "$title",
-       |      "author": "Schindler, Rudolf, 1888-",
-       |      "materialType": {
-       |        "code": "a",
-       |        "value": "Books"
-       |      },
-       |      "bibLevel": {
-       |        "code": "m",
-       |        "value": "MONOGRAPH"
-       |      },
-       |      "publishYear": 1923,
-       |      "catalogDate": "1999-01-01",
-       |      "country": {
-       |        "code": "gw ",
-       |        "name": "Germany"
-       |      }
-       |    }
-    """.stripMargin
 }
